# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Voice-Journal API
  version: 0.3.2
  description: API contract for recording ingestion, uploads, provider configuration, and summaries.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/recordings:
    get:
      summary: List recordings
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Recordings listed
    post:
      summary: Create recording metadata
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordingRequest'
      responses:
        '201':
          description: Recording created
  /api/v1/recordings/{id}:
    get:
      summary: Get recording by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording found
        '404':
          description: Not found
    delete:
      summary: Delete recording and associated resources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording deleted
        '404':
          description: Not found
  /api/v1/recordings/{id}/upload:
    post:
      summary: Upload audio for a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        '201':
          description: Audio uploaded
        '400':
          description: Missing audio
        '404':
          description: Recording not found
  /api/v1/recordings/{id}/file:
    get:
      summary: Download audio file for a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audio file
        '404':
          description: Audio not found
  /api/v1/recordings/{id}/transcribe:
    post:
      summary: Queue transcription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Job queued or deduplicated if already queued/running
  /api/v1/recordings/{id}/speakers:
    put:
      summary: Save speaker labels for transcript speakers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labels]
              properties:
                labels:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '200':
          description: Speaker labels updated
        '404':
          description: Recording not found
  /api/v1/recordings/{id}/tags:
    put:
      summary: Save recording tags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tags]
              properties:
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Tags updated
        '404':
          description: Recording not found
  /api/v1/recordings/{id}/captions:
    put:
      summary: Save caption anchor timestamps for subtitle export
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [anchors]
              properties:
                anchors:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      at:
                        type: number
      responses:
        '200':
          description: Captions updated
        '404':
          description: Recording not found
  /api/v1/recordings/{id}/summaries:
    post:
      summary: Generate summary for recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSummaryRequest'
      responses:
        '201':
          description: Summary generated
        '400':
          description: Bad request
        '404':
          description: Recording not found
  /api/v1/summaries/{id}:
    get:
      summary: Get summary by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Summary found
        '404':
          description: Not found
...
components:
  schemas:
...
    CreateRecordingRequest:
      type: object
      properties:
        title:
          type: string
        metadata:
          type: object
          additionalProperties: true
    CreateSummaryRequest:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          enum: [ollama_local, ollama_cloud, openai, anthropic, deepseek, custom_openai]
        model:
          type: string
        template:
          type: string
          description: Prompt/instruction template used to shape markdown output.
        apiKey:
          type: string
          description: Optional per-request provider API key.
        temperature:
          type: number
        max_tokens:
          type: integer